AMP-49662 Move TS files to 'src' folder
