+++ react-native/typescript/AmpliApp/README.md
@@ -28,7 +28,7 @@ If you want to use your own tracking plan and Ampli SDK
 # Project structure
 * README.md - you are here *
 * src/
-    * [App.tsx](App.tsx) - Example user app using Amply SDK. A good place to start.
+    * [App.tsx](src/App.tsx) - Example user app using Amply SDK. A good place to start.
     * ampli/
-        * [index.ts](ampli/index.ts) - Generated SDK, don't modify by hand. Update with `ampli pull`
-    * [middleware/*](middleware) - Example middlewares for Segment and Event filtering
+        * [index.ts](src/ampli/index.ts) - Generated SDK, don't modify by hand. Update with `ampli pull`
+    * [middleware/*](src/middleware) - Example middlewares for Segment and Event filtering
