+++ react-native/typescript/AmpliApp/src/ampli/index.ts
