+++ react-native/typescript/AmpliApp/src/middleware/loggingMiddleware.ts
