+++ react-native/typescript/AmpliApp/src/middleware/segmentMiddleware.ts
