+++ react-native/typescript/AmpliApp/src/middleware/stopMiddleware.ts
